requirements.txt
capstone
colorama
defusedxml
msgpack
numpy
openpyxl
pefile
phpserialize 
Pillow
pycryptodome
pyelftools
pyperclip
python-magic
xlrd